360MedNet/requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
django-bootstrap3==8.2.2
django-material==0.13.1
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
python3-openid==3.1.0
requests==2.13.0
six==1.10.0
